Project 1 - Hotel Reservation Prediction: Mlflow, Jenkins and GCP Deployment/requirements.txt
pandas
numpy
google-cloud-storage
scikit-learn
pyyaml
task
imbalanced-learn
scipy
joblib
xgboost
mlflow 
